Refactor: Rename matplotlib.py to matplotlib_interface.py
- Rename pyvr/interface/matplotlib.py -> matplotlib_interface.py
- Update import in pyvr/interface/__init__.py
- Update test imports and @patch decorators in test_matplotlib.py

This makes the entry point name more explicit and avoids confusion
with the matplotlib library itself.

All tests pass: 240 tests
