+++ tests/test_interface/test_matplotlib.py
@@ -4,7 +4,7 @@ import pytest
 import numpy as np
 from unittest.mock import Mock, patch, MagicMock
 
-from pyvr.interface.matplotlib import InteractiveVolumeRenderer
+from pyvr.interface.matplotlib_interface import InteractiveVolumeRenderer
 from pyvr.volume import Volume
 from pyvr.datasets import create_sample_volume
 
@@ -16,7 +16,7 @@ def small_volume():
     return Volume(data=data)
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
 def test_interactive_renderer_initialization(mock_renderer_class, small_volume):
     """Test InteractiveVolumeRenderer initializes correctly."""
     # Mock the renderer instance
@@ -32,7 +32,7 @@ def test_interactive_renderer_initialization(mock_renderer_class, small_volume):
     assert len(interface.state.control_points) == 2  # Default control points
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
 def test_custom_dimensions(mock_renderer_class, small_volume):
     """Test custom width and height."""
     mock_renderer = Mock()
@@ -48,7 +48,7 @@ def test_custom_dimensions(mock_renderer_class, small_volume):
     assert call_kwargs['height'] == 768
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
 def test_update_transfer_functions(mock_renderer_class, small_volume):
     """Test transfer function updates."""
     mock_renderer = Mock()
@@ -64,7 +64,7 @@ def test_update_transfer_functions(mock_renderer_class, small_volume):
     mock_renderer.set_transfer_functions.assert_called()
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
 def test_render_volume(mock_renderer_class, small_volume):
     """Test volume rendering."""
     mock_renderer = Mock()
@@ -81,9 +81,9 @@ def test_render_volume(mock_renderer_class, small_volume):
     mock_renderer.render_to_pil.assert_called_once()
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
-@patch('pyvr.interface.matplotlib.GridSpec')
-@patch('pyvr.interface.matplotlib.plt')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.GridSpec')
+@patch('pyvr.interface.matplotlib_interface.plt')
 def test_show_creates_layout(mock_plt, mock_gridspec, mock_renderer_class, small_volume):
     """Test show() creates figure and layout."""
     mock_renderer = Mock()
@@ -111,7 +111,7 @@ def test_show_creates_layout(mock_plt, mock_gridspec, mock_renderer_class, small
     assert interface.color_selector is not None
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
 def test_colormap_change_callback(mock_renderer_class, small_volume):
     """Test colormap change callback."""
     mock_renderer = Mock()
@@ -133,7 +133,7 @@ def test_colormap_change_callback(mock_renderer_class, small_volume):
     interface.image_display.update_image.assert_called_once()
 
 
-@patch('pyvr.interface.matplotlib.VolumeRenderer')
+@patch('pyvr.interface.matplotlib_interface.VolumeRenderer')
 def test_manual_update(mock_renderer_class, small_volume):
     """Test manual update() method."""
     mock_renderer = Mock()
